+++ src/main/java/io/github/randomusert/mods/tincraft_reborn/blocks/UltraCrafter.java
@@ -1,0 +1,49 @@
+package io.github.randomusert.mods.tincraft_reborn.blocks;
+
+import com.mojang.serialization.MapCodec;
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.SimpleMenuProvider;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.ContainerLevelAccess;
+import net.minecraft.world.inventory.CraftingMenu;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import org.jspecify.annotations.Nullable;
+
+public class UltraCrafter extends Block {
+    public static final MapCodec<UltraCrafter> CODEC = simpleCodec(UltraCrafter::new);
+
+    private static final Component CONTAINER_TITLE = Component.translatable("container.tincraft_reborn.ultra_crafting");
+
+    public UltraCrafter(BlockBehaviour.Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    protected MapCodec<? extends Block> codec() {
+        return CODEC;
+    }
+
+    @Override
+    protected InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos, Player player, BlockHitResult hitResult) {
+        if (!level.isClientSide()) {
+            player.openMenu(state.getMenuProvider(level, pos));
+
+        }
+
+        return InteractionResult.PASS;
+    }
+
+    @Override
+    protected @Nullable MenuProvider getMenuProvider(BlockState state, Level level, BlockPos pos) {
+        return new SimpleMenuProvider(
+                (containerId, inv, player) -> new CraftingMenu(containerId, inv, ContainerLevelAccess.create(level, pos)), CONTAINER_TITLE
+        );
+    }
+}