+++ src/main/java/io/github/randomusert/mods/tincraft_reborn/init/TCRMenuTypes.java
@@ -1,0 +1,4 @@
+package io.github.randomusert.mods.tincraft_reborn.init;
+
+public class TCRMenuTypes {
+}