+++ src/main/java/io/github/randomusert/mods/tincraft_reborn/recipe/ultracrafting/UltraCraftingRecipe.java
@@ -1,0 +1,58 @@
+package io.github.randomusert.mods.tincraft_reborn.recipe.ultracrafting;
+
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ItemStackTemplate;
+import net.minecraft.world.item.crafting.*;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class UltraCraftingRecipe implements Recipe<UltraCraftingInput> {
+    private final Recipe.CommonInfo commonInfo;
+    private final UltraCraftingRecipe.BlockBookInfo bookInfo;
+    private final BlockState inputState;
+    private final Ingredient inputItem;
+    private final ItemStackTemplate result;
+
+    public UltraCraftingRecipe(Recipe.CommonInfo commonInfo, UltraCraftingRecipe.BlockBookInfo, ) {}
+
+
+    @Override
+    public boolean matches(UltraCraftingInput ultraCraftingInput, Level level) {
+        return false;
+    }
+
+    @Override
+    public ItemStack assemble(UltraCraftingInput ultraCraftingInput) {
+        return null;
+    }
+
+    @Override
+    public boolean showNotification() {
+        return false;
+    }
+
+    @Override
+    public String group() {
+        return "";
+    }
+
+    @Override
+    public RecipeSerializer<? extends Recipe<UltraCraftingInput>> getSerializer() {
+        return null;
+    }
+
+    @Override
+    public RecipeType<? extends Recipe<UltraCraftingInput>> getType() {
+        return null;
+    }
+
+    @Override
+    public PlacementInfo placementInfo() {
+        return null;
+    }
+
+    @Override
+    public RecipeBookCategory recipeBookCategory() {
+        return null;
+    }
+}